pyproject.toml
[tool.poetry]
name = "top200-solo"
version = "0.1.0"
description = ""
authors = ["Sam Nordlinger <19315410+SNordlinger@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "top200_solo"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-boardgamegeek = {git = "https://github.com/mmonkey/boardgamegeek.git"}
pytest = "^7.4.3"
ply = "^3.11"
thefuzz = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
